+++ web-admin/src/features/organizations/users/CreateUserGroupDialog.svelte
@@ -14,16 +14,44 @@
   import { object, string } from "yup";
   import { page } from "$app/stores";
   import { createAdminServiceListUsergroupMemberUsers } from "@rilldata/web-admin/client";
+  import { eventBus } from "@rilldata/web-common/lib/event-bus/event-bus";
 
   export let open = false;
   export let groupName: string;
-  export let onCreate: (name: string) => void;
 
   $: organization = $page.params.organization;
   $: listUsergroupMemberUsers = createAdminServiceListUsergroupMemberUsers(
     organization,
     groupName,
   );
+
+  const createUserGroup = createAdminServiceCreateUsergroup();
+
+  async function handleCreate(newName: string) {
+    try {
+      await $createUserGroup.mutateAsync({
+        organization: organization,
+        data: {
+          name: newName,
+        },
+      });
+
+      await queryClient.invalidateQueries(
+        getAdminServiceListOrganizationMemberUsergroupsQueryKey(organization),
+      );
+
+      groupName = "";
+      open = false;
+
+      eventBus.emit("notification", { message: "User group created" });
+    } catch (error) {
+      console.error("Error creating user group", error);
+      eventBus.emit("notification", {
+        message: "Error creating user group",
+        type: "error",
+      });
+    }
+  }
 
   const formId = "create-user-group-form";
 
@@ -53,7 +81,7 @@
         const values = form.data;
 
         try {
-          await onCreate(values.name);
+          await handleCreate(values.name);
           open = false;
         } catch (error) {
           console.error(error);
